requirements.txt
Flask==2.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
WTForms==3.1.2
Werkzeug==2.0.3
tensorflow-cpu==2.6.0
keras==2.6.0
opencv-python==4.10.0.84
scikit-image==0.18.3
pillow==10.4.0
matplotlib==3.2.2
reportlab==4.2.5
numpy==1.19.5
pandas==1.3.0
